efficiency fix: calling SkyMap::update() less often.  We were doing a "non-computational update" on each KStarsData::updateTime() call (i.e., 10 times a second).  However, we only need to do one if the LT/UT/ST time changes.  Fixed.
Also, modified timespinbox so that it is wide enough to show the timestep
text.

svn path=/branches/make_kdeedu_cool/kdeedu/kstars/; revision=183941

+++ kstars/viewopsdialog.cpp
@@ -494,7 +494,7 @@ ViewOpsDialog::ViewOpsDialog( QWidget *parent )
 	animateSlewing->setChecked( ksw->options()->useAnimatedSlewing );
 
 	hideSpinBox = new TimeStepBox( AdvancedTab, "HideSpinBox" );
-	hideSpinBox->tsbox()->changeScale( ksw->options()->slewTimeScale );
+	hideSpinBox->tsbox()->changeScale( (float)ksw->options()->slewTimeScale );
 	QLabel *hsbLabel = new QLabel( AdvancedTab, "HSBLabel" );
 	hsbLabel->setText( i18n( "Also hide if time scale greater than:" ) );
